backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
requests==2.31.0
gunicorn==21.2.0
cryptography==41.0.7
pycryptodome==3.19.0
kafka-python==2.0.2
confluent-kafka==2.3.0
# Security & Performance
flask-limiter==3.5.0
flask-compress==1.14
# Logging
python-json-logger==2.0.7
# Validation
marshmallow==3.21.0
# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
# Authentication
PyJWT==2.8.0
bcrypt==4.1.2
# Email/SMS (optional)
python-email-validator==2.1.0
# Note: kafka-python and confluent-kafka are optional - system works with mock data if not installed

